randevu_backend/requirements.txt
# Flask and core extensions
Flask==2.3.3
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

# Security
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1

# CORS and API
Flask-CORS==4.0.0

# Utilities
Werkzeug==2.3.7
email-validator==2.1.0
Pillow==10.0.1  # For image processing
pytz==2023.3.post1  # For timezone handling

# Serving
gunicorn==21.2.0  # For production deployment